+++ lib/providers/points_provider.dart
@@ -4,7 +4,9 @@ import 'package:periodik/utils/collections.dart';
 
 final pointsProvider = StreamProvider.autoDispose.family<List<Point>, String>(
   (ref, signalId) async* {
-    await for (final points in Collections.points(signalId).snapshots()) {
+    await for (final points in Collections.points(
+      signalId,
+    ).orderBy('date').snapshots()) {
       yield points.docs.map((document) {
         return Point.fromJsonDocument(
           document.id,