+++ lib/providers/signals_provider.dart
@@ -3,7 +3,7 @@ import 'package:periodik/models/signal.dart';
 import 'package:periodik/utils/collections.dart';
 
 final signalsProvider = StreamProvider.autoDispose<List<Signal>>((ref) async* {
-  await for (final signals in Collections.signals.snapshots()) {
+  await for (final signals in Collections.signals.orderBy('name').snapshots()) {
     yield signals.docs.map((document) {
       return Signal.fromJsonDocument(
         document.id,